api/requirements.txt
tensorflow==2.12.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api
matplotlib
numpy